Fix boot if crate off error
We can walk for the module numbers but not the indexes if the crate is "off". This results in a python error due to the string parsing for the indexes (slots). So we now detect if the crate is off, ask you to explicitly start it (for safety) and then rerun the startup script when your done. 

i also removed the strip_most set of snmp flags since we dont use that any more