Improvements to the __init__
Moved the lockfile check to a function so that we dont have to maintain two copies in case we find a bug or decide to change the lock file structures.

Did the same thing to the check for IsThisAnMPOD() for the same reasons. Make sure we only have one place to update things if wiener gets another block of MACs or if we decide to add "sys" level calls (like fan speeds or temps for monitor reasons), and then shoehorn in some functions for CRATEs or even abstract it up another layer if needed.